requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.2.6
django-ninja==0.22.2
pydantic==1.10.13
sqlparse==0.4.4
typing-extensions==4.8.0
numpy==1.14.0
scikit-learn==0.19.0
pandas==0.22.0
pyarrow==0.8.0
